Move method removeLast to StringUtils.
